Tests/LibWeb/Text/expected/github-diff-table-fixed-layout-row-height.txt

fixed row height: 24
auto row height: 24
narrow fixed row height: 114
